refactor: 대회일정, 상세, 작성, 수정 페이지 구조 변경 csr -> ssr
